+++ packages/gestalt/src/Label/Label.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import styles from './Label.css';
 
 type Props = {|
-  children?: any,
+  children?: React.Node,
   htmlFor: string,
 |};